backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
Pillow>=10.0
openpyxl>=3.1
reportlab>=4.0
psycopg[binary]>=3.1
gunicorn>=22.0
django-storages[s3]>=1.14
boto3>=1.34
